Decrease Assembler Cost
Add a 1000 resource refund to assemblers to better balance the cost of purchasing a second one. This is rarely ever used late game because it's a far too costly option. Furthermore, the true cost with power grid strain included is actually is plus 2000 consort and 1579 empire. So under new totals its actually 9000 consort and 8579 empire.