Added comments to each player tip (#586)
* Added notes to each item

To help understand each item, i have tried to comment / explain them. 
Also if Vertex could update them as to how they are actually / would be implemented to aid in understanding and hopefully lead to other options. This would not need to be detailed, but to give a better idea of what is actually going to be the trigger for each. For example, on the Grenade Spam, update it with maybe how 'quick succession' is going to be defined in implementation (2 grenades in 2 seconds).