Adds fallback mechanism for offmap fixes
This adds a check and reverts back to the previous ND_OnStructureCreated method when the detour is not available. There is no reason to believe this event would ever occur since signatures are the most resilient amongst gamedata entries and it's unlikely there will be another game update at this point.