Adds capability to process structure requests before they're built (#626)
* Add include file for structure detour plugin

* Add structure detour plugin

* Add custom gamedata file for structure detour plugin

* Minor grammatical fixes

* Integrates offmap fixes plugin with structure detours

* Minor grammatical fixes

* Integrates building fixes plugin with structure detours

* Removes dead code

Removes two code excerpts from the include file duplication for AM webcompiler that are not necessary for compilation of nd_structure_intercept

* Refactoring to use existing ND_Structures enum

* Uses array to index to UTIL_Commander_FailureText

* Removes SetFailState for plugins using the forward

* Creates IsPositionInvalid

Cleans up the code for the detour by putting the logic in a bool IsPositionInvalid function

* Creates functions to check proximity violations

Adds two functions:
bool IsPowerTooCloseToWall(team, position)
bool IsWallTooCloseToPower(team, position)

* Adds fallback mechanism for building fixes

This adds a check and reverts back to the previous ND_OnStructureCreated method when the detour is not available. There is no reason to believe this event would ever occur since signatures are the most resilient amongst gamedata entries and it's unlikely there will be another game update at this point.

* Adds fallback mechanism for offmap fixes

This adds a check and reverts back to the previous ND_OnStructureCreated method when the detour is not available. There is no reason to believe this event would ever occur since signatures are the most resilient amongst gamedata entries and it's unlikely there will be another game update at this point.

* Timer best practices implemented

* Adds GetStructureDisplayName

Adds a stock to get a friendly display name based on the ND_Structure enum provided. Optional parameter is provided to allow to return the name in all upper-case.

* Switches to GetStructureDisplayName for building_fixes

Switches to the GetStructureDisplayName to generate the notice to the commander

* Changes to StrEqual

* Removing nd_building_fixes from wrong directory

* Switches to StrEqual for building_fixes

Switches to using StrEqual in ND_OnStructureCreated code

* Clarifying error message for detour unavailable

* Switches to array for GetStructureDisplayName

* Removes unnecessary include

* Remove web compiler support

Co-authored-by: stickz <stickman002@mail.com>